conversão do jobs para ts
